machine_learning/requirements.txt
pandas
sf-hamilton
scikit-learn
click
dagworks-sdk